requirements.txt
aws-cdk-lib==2.121.1
cdk-nag==2.28.12
cdk-opinionated-constructs==3.3.9
constructs==10.3.0
docutils==0.20.1
email-validator==2.1.0.post1
pydantic==2.5.3
pydantic-core>=2.14
pydoctor==23.9.1
pyyaml==6.0.1
